+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
 ## What's Changed:
-* 
+* Removed Broken AprilFools Patch
 
 **Full Changelog**: [4.2.2...4.2.3](<https://github.com/LavaGang/MelonLoader.Installer/compare/4.2.2...4.2.3>)